requirements.txt
pycryptodome>=3.19.1
tqdm>=4.66.1

